Added an ID variable to the Post class
This isn't passed to the constructor, but calculated once the all the
posts have been loaded, it is the total number of posts subtract the key
value, resulting in a static id, which should never change. The keys are
applied after the sorting has occurred to ensure that the ids correspond
to the sorted post order. A Time variable has also been added and is
used when ordering the posts.