Changed config structure, fixed post ordering
The config relevant to the posts is now held statically within the Post
class with default values, this means that the user does not have to
override every option. The date format has been changed to dd-mm-yyyy as
this matches a European time within PHP. A choice of input types will be
offered in future.