Added post navigation text override
This allows the user to pass their own text, which will be displayed
instead of the default Previous and Next links. This also accepts the
arguments title, date, author and category (surrounded with percentage
symbols, e.g. %title%), which will be replaced by the appropriate
attributes from the previous or next posts. Defaults have also been
added to the configuration. Post navigation can now be enabled or
disabled.