Added the ability for sticky posts
A sticky post is displayed at the top of the first page regardless of
its date, this is useful for example if you have an important article
that you want to have displayed first always. There is an override that
can be passed, that forces a normal chronological order, this is for use
with queries so sticky posts are not forced to the top.